5758 - Split tests
